+++ CHANGES.rst
@@ -16,8 +16,9 @@ Changelog
 - Add ``Predicate``.
 - Add ``String``.
 - Add ``Tuple``.
-- Make ``Except`` **not** call *value* if it's callable. (**breaking change**)
-- Make ``Except`` return ``True`` if all assertions pass.
+- Rename ``Except`` to ``except``. (**breaking change**)
+- Make ``except`` **not** call `value` if it's callable. (**breaking change**)
+- Make ``except`` return ``True`` if all assertions pass.
 
 
 v0.0.1 (2015-05-07)